Added function to check for output lines in any order
See editorconfig/editorconfig#375.  NOTE: requires CMake 3.1.0+.

New function ec_test_lines() will test for up to three lines in any
order.  The limit of three lines is because CMake is limited to nine
groups in a single regex (as far as I know), and each line requires one
group in this implementation.